Execute the test only on JVM
