+++ tests/shared/src/test/scala/cats/effect/IOPropSpec.scala
@@ -20,7 +20,7 @@ import cats.syntax.all._
 import cats.effect.implicits._
 import cats.effect.std.Queue
 
-import org.scalacheck.Arbitrary, Arbitrary.arbitrary
+import org.scalacheck.Arbitrary.arbitrary
 
 import org.specs2.ScalaCheck
 
@@ -79,23 +79,6 @@ class IOPropSpec extends IOPlatformSpecification with Discipline with ScalaCheck
           l.map(IO.pure(_)).parSequence.flatMap { expected =>
             l.map(IO.pure(_)).parSequenceN(n).mustEqual(expected)
           }
-      }
-
-    }
-
-    "evaluate fibers correctly in presence of a parasitic execution context" in real {
-      implicit val ticker = Ticker()
-
-      val test = IO(implicitly[Arbitrary[IO[Int]]].arbitrary.sample.get).flatMap { io =>
-        IO.delay(io.eqv(io))
-      }
-
-      val iterations = 5000
-
-      List.fill(iterations)(test).sequence.map(_.count(identity)).flatMap { c =>
-        IO {
-          c mustEqual iterations
-        }
       }
     }
   }
